fix: Bug #1, bump version to v0.2.0
Also, update README.md.

The problem was my `pp` arithmetic was wrong, I was using `layout.size()`
which is the size of the entire allocation. The fix is to use the size
of a Protocol message. So I added `protocol_layout` and now use
`protocol_layout.size()` to compute an address for available Protocol
messages.

Resolves #1