feat: Use OnceCell sync Lazy for MyAllocator
The Box::new_in needs to be passed a pure Trait not an Object Trait
becuase Box::new_in takes ownership and consumes an Object Trait.

So now MyAllocator is a pure Trait as there are no fields. So now we
can allocate more than one without having to "clone" MyAllocator.